requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-JWT-Extended==4.1.0
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.1.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
virtualenv==20.4.3
Werkzeug==1.0.1
